updateHeader -> render; fetch error handling -> product-api
