+++ src/js/helpers.js
@@ -1,12 +1,13 @@
 //Допоміжні функції
 
-import { refs } from './refs';
-import { LS } from './storage';
+// import { refs } from './refs';
+// import { LS } from './storage';
 import { getCategories, getProducts } from './products-api';
 import {
   renderCategories,
   renderProducts,
   clearProducts,
+  updateHeader,
 } from './render-function';
 
 export const productsPerPage = 12;
@@ -16,22 +17,18 @@ let searchQuery = null;
 
 export async function loadPage() {
   updateHeader();
-  try {
-    const categories = await getCategories();
+  const categories = await getCategories();
+  if (categories) {
     renderCategories(['all', ...categories]);
-  } catch (error) {
-    console.log(error);
+    handleProducts();
   }
-  handleProducts();
 }
 
 async function handleProducts() {
-  try {
-    const products = await getProducts(page, category, searchQuery);
+  const products = await getProducts(page, category, searchQuery);
+  if (products) {
     renderProducts(products, page, searchQuery);
     searchQuery = null;
-  } catch (error) {
-    console.log(error);
   }
 }
 
@@ -52,17 +49,4 @@ export function applyFilter(categorySelected, searchFormQuery) {
   clearProducts(category, categorySelected);
   category = categorySelected;
   handleProducts();
-}
-
-export function updateHeader() {
-  let cart = 0;
-  let wish = 0;
-  LS.getKeys().forEach(item => {
-    console.log(item);
-    item.qty && cart++;
-    item.wish && wish++;
-  });
-  console.log(cart, wish);
-  refs.body.querySelector('[data-cart-count]').textContent = cart;
-  refs.body.querySelector('[data-wishlist-count]').textContent = wish;
 }
